+++ NEWS.md
@@ -1,6 +1,6 @@
 # stockfish (development version)
 
-* Support for M1 Mac compilation.
+* Generalized tests on M1 Mac.
 * Updated code for full Solaris compilation.
 
 # stockfish 0.1.2