+++ NEWS.md
@@ -1,4 +1,4 @@
-# stockfish (development version)
+# stockfish 0.1.3
 
 * Generalized tests on M1 Mac.
 * Updated code for full Solaris compilation.